requirements.txt
google-genai>=0.1.0
python-dotenv>=1.0.0
Flask>=2.0.0
gunicorn>=20.1.0
